+++ src/Elements/Typography.php
@@ -106,6 +106,11 @@ class Typography extends \Bricks\Element {
 			'title' => esc_html__( 'Item Styling', 'bricks-style-guide' ),
 			'tab'   => 'content',
 		];
+
+		$this->control_groups['typographyOverride'] = [
+			'title' => esc_html__( 'Typography Override', 'bricks-style-guide' ),
+			'tab'   => 'content',
+		];
 	}
 
 	/**
@@ -155,6 +160,60 @@ class Typography extends \Bricks\Element {
 			'css'         => [
 				[
 					'property' => 'gap',
+					'selector' => '',
+				],
+			],
+		];
+
+		$this->controls['flexDirection'] = [
+			'group'   => 'layout',
+			'label'   => esc_html__( 'Direction', 'bricks-style-guide' ),
+			'type'    => 'direction',
+			'css'     => [
+				[
+					'property' => 'flex-direction',
+					'selector' => '',
+				],
+			],
+		];
+
+		$this->controls['flexWrap'] = [
+			'group'   => 'layout',
+			'label'   => esc_html__( 'Wrap', 'bricks-style-guide' ),
+			'type'    => 'select',
+			'options' => [
+				'nowrap'       => esc_html__( 'No Wrap', 'bricks-style-guide' ),
+				'wrap'         => esc_html__( 'Wrap', 'bricks-style-guide' ),
+				'wrap-reverse' => esc_html__( 'Wrap Reverse', 'bricks-style-guide' ),
+			],
+			'inline'  => true,
+			'css'     => [
+				[
+					'property' => 'flex-wrap',
+					'selector' => '',
+				],
+			],
+		];
+
+		$this->controls['alignItems'] = [
+			'group'   => 'layout',
+			'label'   => esc_html__( 'Align Items', 'bricks-style-guide' ),
+			'type'    => 'align-items',
+			'css'     => [
+				[
+					'property' => 'align-items',
+					'selector' => '',
+				],
+			],
+		];
+
+		$this->controls['justifyContent'] = [
+			'group'   => 'layout',
+			'label'   => esc_html__( 'Justify Content', 'bricks-style-guide' ),
+			'type'    => 'justify-content',
+			'css'     => [
+				[
+					'property' => 'justify-content',
 					'selector' => '',
 				],
 			],
@@ -349,6 +408,53 @@ class Typography extends \Bricks\Element {
 					'selector' => '.bsg-typography-item',
 				],
 			],
+		];
+
+		// Typography Override controls.
+		$this->controls['overrideChildTypography'] = [
+			'group'       => 'typographyOverride',
+			'label'       => esc_html__( 'Override Child Typography Settings', 'bricks-style-guide' ),
+			'type'        => 'checkbox',
+			'description' => esc_html__( 'Enable to control typography settings for all child items from here.', 'bricks-style-guide' ),
+		];
+
+		$this->controls['overrideLabelTypography'] = [
+			'group'    => 'typographyOverride',
+			'label'    => esc_html__( 'Label Typography', 'bricks-style-guide' ),
+			'type'     => 'typography',
+			'css'      => [
+				[
+					'property' => 'font',
+					'selector' => '.bsg-typography-item__label',
+				],
+			],
+			'required' => [ 'overrideChildTypography', '!=', '' ],
+		];
+
+		$this->controls['overrideMetaTypography'] = [
+			'group'    => 'typographyOverride',
+			'label'    => esc_html__( 'Meta Typography', 'bricks-style-guide' ),
+			'type'     => 'typography',
+			'css'      => [
+				[
+					'property' => 'font',
+					'selector' => '.bsg-typography-item__meta',
+				],
+			],
+			'required' => [ 'overrideChildTypography', '!=', '' ],
+		];
+
+		$this->controls['overrideMetaLabelTypography'] = [
+			'group'    => 'typographyOverride',
+			'label'    => esc_html__( 'Meta Label Typography', 'bricks-style-guide' ),
+			'type'     => 'typography',
+			'css'      => [
+				[
+					'property' => 'font',
+					'selector' => '.bsg-typography-item__meta-label',
+				],
+			],
+			'required' => [ 'overrideChildTypography', '!=', '' ],
 		];
 	}